+++ climpred/reference.py
@@ -1,4 +1,5 @@
 import inspect
+import warnings
 
 import pandas as pd
 import xarray as xr
@@ -158,7 +159,7 @@ def _adapt_member_for_reference_forecast(lforecast, lverif, metric, comparison, 
 
 
 def compute_climatology(
-    hind,
+    initialized,
     verif=None,
     metric="pearson_r",
     comparison="m2e",
@@ -169,8 +170,8 @@ def compute_climatology(
     """Computes the skill of a climatology forecast.
 
     Args:
-        hind (xarray object): The initialized ensemble.
-        verif (xarray object): control data, not needed
+        initialized (xarray.Dataset): The initialized ensemble.
+        verif (xarray.Dataset): control data, not needed
         metric (str): Metric name to apply at each lag for the persistence computation.
             Default: 'pearson_r'
         dim (str or list of str): dimension to apply metric over.
@@ -178,14 +179,14 @@ def compute_climatology(
             (see the arguments required for a given metric in :ref:`Metrics`).
 
     Returns:
-        clim (xarray object): Results of climatology forecast with the input metric
+        clim (xarray.Dataset): Results of climatology forecast with the input metric
             applied.
     """
     seasonality_str = OPTIONS["seasonality"]
 
     if isinstance(dim, str):
         dim = [dim]
-    has_valid_lead_units(hind)
+    has_valid_lead_units(initialized)
 
     # get metric/comparison function name, not the alias
     metric = METRIC_ALIASES.get(metric, metric)
@@ -194,8 +195,8 @@ def compute_climatology(
     comparison = get_comparison_class(comparison, ALL_COMPARISONS)
     metric = get_metric_class(metric, ALL_METRICS)
 
-    if "iteration" in hind.dims:
-        hind = hind.isel(iteration=0, drop=True)
+    if "iteration" in initialized.dims:
+        initialized = initialized.isel(iteration=0, drop=True)
 
     if comparison.hindcast:
         kind = "hindcast"
@@ -203,10 +204,10 @@ def compute_climatology(
         kind = "perfect"
 
     if kind == "perfect":
-        forecast, verif = comparison.function(hind, metric=metric)
+        forecast, verif = comparison.function(initialized, metric=metric)
         climatology_day = verif.groupby(f"init.{seasonality_str}").mean()
     else:
-        forecast, verif = comparison.function(hind, verif, metric=metric)
+        forecast, verif = comparison.function(initialized, verif, metric=metric)
         climatology_day = verif.groupby(f"time.{seasonality_str}").mean()
 
     climatology_day_forecast = climatology_day.sel(
@@ -233,11 +234,10 @@ def compute_climatology(
 
 @is_xarray([0, 1])
 def compute_persistence(
-    hind,
+    initialized,
     verif,
     metric="pearson_r",
     alignment="same_verifs",
-    add_attrs=True,
     dim="init",
     comparison="m2o",
     **metric_kwargs,
@@ -245,12 +245,13 @@ def compute_persistence(
     """Computes the skill of a persistence forecast from a simulation.
 
     Args:
-        hind (xarray object): The initialized ensemble.
-        verif (xarray object): Verification data.
+        initialized (xarray.Dataset): The initialized ensemble.
+        verif (xarray.Dataset): Verification data.
         metric (str): Metric name to apply at each lag for the persistence computation.
             Default: 'pearson_r'
         alignment (str): which inits or verification times should be aligned?
-            - maximize/None: maximize the degrees of freedom by slicing ``hind`` and
+
+            - maximize/None: maximize the degrees of freedom by slicing ``initialized`` and
             ``verif`` to a common time frame at each lead.
             - same_inits: slice to a common init frame prior to computing
             metric. This philosophy follows the thought that each lead should be based
@@ -258,14 +259,13 @@ def compute_persistence(
             - same_verif: slice to a common/consistent verification time frame prior to
             computing metric. This philosophy follows the thought that each lead
             should be based on the same set of verification dates.
+
         dim (str or list of str): dimension to apply metric over.
-        add_attrs (bool): write climpred compute_persistence args to attrs.
-            default: True
         ** metric_kwargs (dict): additional keywords to be passed to metric
             (see the arguments required for a given metric in :ref:`Metrics`).
 
     Returns:
-        pers (xarray object): Results of persistence forecast with the input metric
+        pers (xarray.Dataset): Results of persistence forecast with the input metric
             applied.
 
     Reference:
@@ -273,15 +273,17 @@ def compute_persistence(
           Empirical methods in short-term climate prediction.
           Oxford University Press, 2007.
 
+    See also:
+        * :py:func:`~climpred.reference.compute_persistence`
     """
     if isinstance(dim, str):
         dim = [dim]
     # Check that init is int, cftime, or datetime; convert ints or cftime to datetime.
-    hind = convert_time_index(hind, "init", "hind[init]")
+    initialized = convert_time_index(initialized, "init", "initialized[init]")
     verif = convert_time_index(verif, "time", "verif[time]")
     # Put this after `convert_time_index` since it assigns 'years' attribute if the
     # `init` dimension is a `float` or `int`.
-    has_valid_lead_units(hind)
+    has_valid_lead_units(initialized)
 
     # get metric/comparison function name, not the alias
     metric = METRIC_ALIASES.get(metric, metric)
@@ -293,24 +295,26 @@ def compute_persistence(
     metric = get_metric_class(metric, ALL_METRICS)
     # If lead 0, need to make modifications to get proper persistence, since persistence
     # at lead 0 is == 1.
-    if [0] in hind.lead.values:
-        hind = hind.copy()
+    if [0] in initialized.lead.values:
+        initialized = initialized.copy()
         with xr.set_options(keep_attrs=True):  # keeps lead.attrs['units']
-            hind["lead"] = hind["lead"] + 1
-        n, freq = get_lead_cftime_shift_args(hind.lead.attrs["units"], 1)
+            initialized["lead"] = initialized["lead"] + 1
+        n, freq = get_lead_cftime_shift_args(initialized.lead.attrs["units"], 1)
         # Shift backwards shift for lead zero.
-        hind["init"] = shift_cftime_index(hind, "init", -1 * n, freq)
+        initialized["init"] = shift_cftime_index(initialized, "init", -1 * n, freq)
     # temporarily change `init` to `time` for comparison to verification data time.
-    hind = hind.rename({"init": "time"})
+    initialized = initialized.rename({"init": "time"})
 
-    inits, verif_dates = return_inits_and_verif_dates(hind, verif, alignment=alignment)
+    inits, verif_dates = return_inits_and_verif_dates(
+        initialized, verif, alignment=alignment
+    )
 
     if metric.normalize:
         metric_kwargs["comparison"] = __e2c
-    dim = _rename_dim(dim, hind, verif)
+    dim = _rename_dim(dim, initialized, verif)
 
     plag = []
-    for i in hind.lead.values:
+    for i in initialized.lead.values:
         lforecast = verif.sel(time=inits[i])
         lverif = verif.sel(time=verif_dates[i])
         lforecast, dim = _adapt_member_for_reference_forecast(
@@ -322,13 +326,131 @@ def compute_persistence(
     pers = xr.concat(plag, "lead")
     if "time" in pers:
         pers = pers.dropna(dim="time").rename({"time": "init"})
-    pers["lead"] = hind.lead.values
+    pers["lead"] = initialized.lead.values
     return pers
+
+
+def compute_persistence_from_first_lead(
+    initialized,
+    verif,
+    metric="pearson_r",
+    alignment="same_inits",
+    dim="init",
+    comparison="m2e",
+    **metric_kwargs,
+):
+    """Computes the skill of a persistence forecast based on the first lead available
+    in the initialized dataset. This function unlike ``compute_persistence`` is
+    sensitive to ``comparison``. Requires
+    ``climpred.set_options(PerfectModel_persistence_from_initialized_lead_0=True)``.
+
+    Args:
+        initialized (xarray.Dataset): The initialized ensemble.
+        verif (xarray.Dataset): Verification data. Not used.
+        metric (str): Metric name to apply at each lag for the persistence computation.
+            Default: 'pearson_r'
+        alignment (str): which inits or verification times should be aligned?
+
+            - ``maximize``: maximize the degrees of freedom by slicing ``initialized`` and
+              ``verif`` to a common time frame at each lead.
+            - ``same_inits``: slice to a common init frame prior to computing
+              metric. This philosophy follows the thought that each lead should be based
+              on the same set of initializations.
+            - ``same_verif``: slice to a common/consistent verification time frame prior to
+              computing metric. This philosophy follows the thought that each lead
+              should be based on the same set of verification dates.
+
+        dim (str or list of str): dimension to apply metric over.
+        ** metric_kwargs (dict): additional keywords to be passed to metric
+            (see the arguments required for a given metric in :ref:`Metrics`).
+
+    Returns:
+        pers (xarray.Dataset): Results of persistence forecast with the input metric
+            applied.
+
+    Example:
+        >>> with climpred.set_options(PerfectModel_persistence_from_initialized_lead_0=True):
+        ...     PerfectModelEnsemble.verify(metric="mse", comparison="m2e",
+        ...         dim=["init", "member"], reference="persistence"
+        ...     ).sel(skill='persistence')  # persistence sensitive to comparison
+        <xarray.Dataset>
+        Dimensions:  (lead: 20)
+        Coordinates:
+          * lead     (lead) int64 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20
+            skill    <U11 'persistence'
+        Data variables:
+            tos      (lead) float32 0.01056 0.01962 0.02925 ... 0.08033 0.08731 0.07578
+        Attributes:
+            prediction_skill_software:                         climpred https://clim...
+            skill_calculated_by_function:                      PerfectModelEnsemble....
+            number_of_initializations:                         12
+            number_of_members:                                 10
+            metric:                                            mse
+            comparison:                                        m2e
+            dim:                                               ['init', 'member']
+            reference:                                         ['persistence']
+            PerfectModel_persistence_from_initialized_lead_0:  True
+
+
+        >>> with climpred.set_options(PerfectModel_persistence_from_initialized_lead_0=False):
+        ...     PerfectModelEnsemble.verify(metric="mse", comparison="m2e",
+        ...         dim=["init", "member"], reference="persistence"
+        ...     ).sel(skill='persistence')  # persistence not sensitive to comparison
+        <xarray.Dataset>
+        Dimensions:  (lead: 20)
+        Coordinates:
+          * lead     (lead) int64 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20
+            skill    <U11 'persistence'
+        Data variables:
+            tos      (lead) float32 0.02794 0.04554 0.08024 ... 0.06327 0.09077 0.05898
+        Attributes:
+            prediction_skill_software:                         climpred https://clim...
+            skill_calculated_by_function:                      PerfectModelEnsemble....
+            number_of_initializations:                         12
+            number_of_members:                                 10
+            metric:                                            mse
+            comparison:                                        m2e
+            dim:                                               ['init', 'member']
+            reference:                                         ['persistence']
+            PerfectModel_persistence_from_initialized_lead_0:  False
+
+    Reference:
+        * Chapter 8 (Short-Term Climate Prediction) in Van den Dool, Huug.
+          Empirical methods in short-term climate prediction.
+          Oxford University Press, 2007.
+
+    See also:
+        * :py:func:`~climpred.reference.compute_persistence`
+
+    """
+    if isinstance(dim, str):
+        dim = [dim]
+    # Check that init is int, cftime, or datetime; convert ints or cftime to datetime.
+    initialized = convert_time_index(initialized, "init", "initialized[init]")
+    has_valid_lead_units(initialized)
+
+    # get metric/comparison function name, not the alias
+    metric = METRIC_ALIASES.get(metric, metric)
+    comparison = COMPARISON_ALIASES.get(comparison, comparison)
+
+    comparison = get_comparison_class(comparison, ALL_COMPARISONS)
+    metric = get_metric_class(metric, ALL_METRICS)
+
+    forecast, observations = comparison.function(initialized, metric=metric)
+    forecast, dim = _adapt_member_for_reference_forecast(
+        forecast, observations, metric, comparison, dim
+    )
+    # persistence is forecast lead 0
+    persistence_skill = metric.function(
+        forecast.isel(lead=0, drop=True), observations, dim=dim, **metric_kwargs
+    )
+    persistence_skill["lead"] = persistence_skill.lead.values
+    return persistence_skill
 
 
 @is_xarray([0, 1])
 def compute_uninitialized(
-    hind,
+    initialized,
     uninit,
     verif,
     metric="pearson_r",
@@ -344,9 +466,9 @@ def compute_uninitialized(
         first lag and then projected out to any further lags being analyzed.
 
     Args:
-        hind (xarray object): Initialized ensemble.
-        uninit (xarray object): Uninitialized ensemble.
-        verif (xarray object): Verification data with some temporal overlap with the
+        initialized (xarray.Dataset): Initialized ensemble.
+        uninit (xarray.Dataset): Uninitialized ensemble.
+        verif (xarray.Dataset): Verification data with some temporal overlap with the
             uninitialized ensemble.
         metric (str):
             Metric used in comparing the uninitialized ensemble with the verification
@@ -357,7 +479,8 @@ def compute_uninitialized(
                 * m2o : each member to the verification data
         dim (str or list of str): dimension to apply metric over.
         alignment (str): which inits or verification times should be aligned?
-            - maximize/None: maximize the degrees of freedom by slicing ``hind`` and
+
+            - maximize/None: maximize the degrees of freedom by slicing ``initialized`` and
             ``verif`` to a common time frame at each lead.
             - same_inits: slice to a common init frame prior to computing
             metric. This philosophy follows the thought that each lead should be based
@@ -365,19 +488,20 @@ def compute_uninitialized(
             - same_verif: slice to a common/consistent verification time frame prior to
             computing metric. This philosophy follows the thought that each lead
             should be based on the same set of verification dates.
+
         ** metric_kwargs (dict): additional keywords to be passed to metric
 
     Returns:
-        u (xarray object): Results from comparison at the first lag.
+        u (xarray.Dataset): Results from comparison at the first lag.
 
     """
     if isinstance(dim, str):
         dim = [dim]
     # Check that init is int, cftime, or datetime; convert ints or cftime to datetime.
-    hind = convert_time_index(hind, "init", "hind[init]")
+    initialized = convert_time_index(initialized, "init", "initialized[init]")
     uninit = convert_time_index(uninit, "time", "uninit[time]")
     verif = convert_time_index(verif, "time", "verif[time]")
-    has_valid_lead_units(hind)
+    has_valid_lead_units(initialized)
 
     # get metric/comparison function name, not the alias
     metric = METRIC_ALIASES.get(metric, metric)
@@ -387,9 +511,11 @@ def compute_uninitialized(
     metric = get_metric_class(metric, DETERMINISTIC_HINDCAST_METRICS)
     forecast, verif = comparison.function(uninit, verif, metric=metric)
 
-    hind = hind.rename({"init": "time"})
+    initialized = initialized.rename({"init": "time"})
 
-    _, verif_dates = return_inits_and_verif_dates(hind, verif, alignment=alignment)
+    _, verif_dates = return_inits_and_verif_dates(
+        initialized, verif, alignment=alignment
+    )
 
     if metric.normalize:
         metric_kwargs["comparison"] = comparison
@@ -397,7 +523,7 @@ def compute_uninitialized(
     plag = []
     # TODO: `same_verifs` does not need to go through the loop, since it's a fixed
     # skill over all leads
-    for i in hind["lead"].values:
+    for i in initialized["lead"].values:
         # Ensure that the uninitialized reference has all of the
         # dates for alignment.
         dates = list(set(forecast["time"].values) & set(verif_dates[i]))
@@ -410,5 +536,5 @@ def compute_uninitialized(
         # comparison expected for normalized metrics
         plag.append(metric.function(lforecast, lverif, dim=dim, **metric_kwargs))
     uninit_skill = xr.concat(plag, "lead")
-    uninit_skill["lead"] = hind.lead.values
+    uninit_skill["lead"] = initialized.lead.values
     return uninit_skill