+++ CHANGELOG.rst
@@ -2,6 +2,16 @@
 What's New
 ==========
 
+.. ipython:: python
+    :suppress:
+
+    import climpred
+    from climpred import HindcastEnsemble
+    import matplotlib as mpl
+    mpl.rcdefaults()
+    mpl.use('Agg')
+    # cut border when saving (for maps)
+    mpl.rcParams["savefig.bbox"] = "tight"
 
 climpred unreleased (202x-xx-xx)
 ================================
@@ -22,9 +32,8 @@ New Features
 
   .. code-block:: python
 
-      >>> import climpred
       >>> hind = climpred.tutorial.load_dataset('CESM-DP-SST')
-      >>> hind.lead.attrs['units'] = 'years'
+      >>> hind.lead.attrs["units"] = "years"
       >>> climpred.HindcastEnsemble(hind).get_initialized()
       <xarray.Dataset>
       Dimensions:     (lead: 10, member: 10, init: 64)
@@ -75,20 +84,36 @@ New Features
       >>> skill.sst.plot(hue="model", col="month", col_wrap=3)
 
   (:issue:`635`, :pr:`690`) `Aaron Spring`_.
-- :py:meth:`~climpred.classes.HindcastEnsemble.plot_alignment` shows how forecast and
-  observations are aligned based on the `alignment <alignment.html>`_ keyword.
-  This may help understanding which dates are matched for the different ``alignment``
-  approaches. (:issue:`701`, :pr:`702`) `Aaron Spring`_.
+- :py:meth:`~climpred.classes.HindcastEnsemble.plot_alignment` shows how forecast and observations are
+  aligned based on the `alignment <alignment.html>`_ keyword. This may help
+  understanding which dates are matched for the different ``alignment`` approaches.
+  (:issue:`701`, :pr:`702`) `Aaron Spring`_.
+
+  .. ipython:: python
+      :okwarning:
+
+      from climpred.tutorial import load_dataset
+      hindcast = climpred.HindcastEnsemble(load_dataset("CESM-DP-SST")).add_observations(load_dataset("ERSST"))
+      @savefig plotting_MEOW.png width=100%
+      hindcast.plot_alignment(edgecolor="w")
+
 - Add ``attrs`` to new ``coordinates`` created by ``climpred``.
   (:issue:`695`, :pr:`697`) `Aaron Spring`_.
 - Add ``seasonality="weekofyear"`` in ``reference="climatology"``.
   (:pr:`703`) `Aaron Spring`_.
+- Compute ``reference="persistence"`` in
+  :py:class:`~climpred.classes.PerfectModelEnsemble` from ``initialized`` first ``lead``
+  if :py:class:`~climpred.options.set_options`
+  ``(PerfectModel_persistence_from_initialized_lead_0=True)`` (``False`` by default)
+  using :py:func:`~climpred.reference.compute_persistence_from_first_lead`.
+  (:issue:`637`, :pr:`706`) `Aaron Spring`_.
 
 
 Internals/Minor Fixes
 ---------------------
-- Reduce dependencies (:pr:`686`) `Aaron Spring`_.
-- Add `typing <https://docs.python.org/3/library/typing.html>`_ (:issue:`685`, :pr:`692`) `Aaron Spring`_.
+- Reduce dependencies. (:pr:`686`) `Aaron Spring`_.
+- Add `typing <https://docs.python.org/3/library/typing.html>`_.
+  (:issue:`685`, :pr:`692`) `Aaron Spring`_.
 - refactor ``add_attrs`` into :py:meth:`~climpred.classes.HindcastEnsemble.verify` and
   :py:meth:`~climpred.classes.HindcastEnsemble.bootstrap`. Now all keywords are
   captured in the skill dataset attributes ``.attrs``.